Checking for CERT_PATH directory in the fetch-certs script to prevent errors.

+++ scripts/fetch-certs.sh
@@ -8,6 +8,9 @@ CERT_REFERENCE="${3:-main}"
 #-------------------------------------------------------------------------------
 
 echo "> Fetching Zimagi development SSL certificates"
-find "$CERT_PATH" -type d -exec rm -Rf {} \;
-find "$CERT_PATH" -type f -exec rm -f {} \;
+if [ -d "$CERT_PATH" ];
+then
+    find "$CERT_PATH" -type d -exec rm -Rf {} \;
+    find "$CERT_PATH" -type f -exec rm -f {} \;
+fi
 git clone --branch="$CERT_REFERENCE" "$CERT_REPO" "$CERT_PATH"
